requirements.txt
django
djangorestframework
Pillow
requests
psycopg2